requirements.txt
fastapi
uvicorn
git+https://github.com/openai/whisper.git
openai
pdfkit
python-multipart
python-dotenv